debug: add logic to write out any writes to heap
this requires a memory bp to be set before it starts logging. I was just mem bp'ing the RNG seed but presumably other things would work.

im committing this logic because it's something I consider doing from time to time to figure out what memory locations are being written to so it might be useful in the future